+++ src/main/java/ssafy/retrip/api/service/openai/OpenAiClient.java
@@ -2,13 +2,16 @@ package ssafy.retrip.api.service.openai;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.CompletableFuture;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClient;
 
@@ -29,18 +32,20 @@ public class OpenAiClient {
   @Value("${openai.model}")
   private String model;
 
-  /**
-   * GPT-4o에 이미지와 프롬프트를 전송하여 분석 결과를 JSON 형태로 반환
-   */
-  public String analyzeImages(String prompt, List<String> imageDataUrls) {
+  @Async("openAiExecutor")
+  public CompletableFuture<String> analyzeImages(String prompt, List<String> imageDataUrls) {
+    long startTime = System.currentTimeMillis();
+    String threadName = Thread.currentThread().getName();
+
+    log.info("OpenAI API 요청 시작 - Thread: {}, 시작시간: {}",
+        threadName, LocalDateTime.now());
+
     try {
-      // 메시지 구성
       Map<String, Object> message = Map.of(
           "role", "user",
           "content", buildContent(prompt, imageDataUrls)
       );
 
-      // API 요청 본문 구성
       Map<String, Object> requestBody = Map.of(
           "model", model,
           "messages", List.of(message),
@@ -48,7 +53,6 @@ public class OpenAiClient {
           "response_format", Map.of("type", "json_object")
       );
 
-      // API 호출
       String response = restClient.post()
           .uri(apiUrl + "/chat/completions")
           .header(HttpHeaders.AUTHORIZATION, "Bearer " + apiKey)
@@ -57,26 +61,32 @@ public class OpenAiClient {
           .retrieve()
           .body(String.class);
 
-      // 응답에서 콘텐츠 추출
       JsonNode jsonResponse = objectMapper.readTree(response);
-      return jsonResponse.path("choices").path(0).path("message").path("content").asText();
+      long endTime = System.currentTimeMillis();
+      long executionTime = endTime - startTime;
+
+      log.info("OpenAI API 요청 완료 - Thread: {}, 종료시간: {}, 실행시간: {}ms",
+          threadName, LocalDateTime.now(), executionTime);
+
+      String result = jsonResponse.path("choices").path(0).path("message").path("content").asText();
+
+      return CompletableFuture.completedFuture(result);
+
     } catch (Exception e) {
-      log.error("OpenAI API 호출 중 오류 발생", e);
-      throw new RuntimeException("이미지 분석 중 오류가 발생했습니다: " + e.getMessage());
+      log.error("[{}] GPT 호출 실패", threadName, e);
+      return CompletableFuture.failedFuture(
+          new RuntimeException("GPT API 호출 중 오류가 발생했습니다: " + e.getMessage(), e)
+      );
     }
   }
 
-  /**
-   * OpenAI API 요청에 필요한 콘텐츠 구성
-   */
+
   private List<Map<String, Object>> buildContent(String prompt, List<String> imageDataUrls) {
-    // 텍스트 프롬프트
     Map<String, Object> textContent = Map.of(
         "type", "text",
         "text", prompt
     );
 
-    // 이미지 콘텐츠 구성
     List<Map<String, Object>> contents = imageDataUrls.stream()
         .map(url -> Map.of(
             "type", "image_url",
@@ -84,7 +94,6 @@ public class OpenAiClient {
         ))
         .collect(java.util.stream.Collectors.toList());
 
-    // 텍스트 프롬프트를 맨 앞에 추가
     contents.add(0, textContent);
     return contents;
   }
